% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftf.R
\name{createDerivedFtF}
\alias{createDerivedFtF}
\title{Creates useful derived variables and returns dt}
\usage{
createDerivedFtF(dt)
}
\description{
\code{createDerivedFtF} sets proper dates and times for R
}
\details{
adds new derived variables to the passed data.table and returns it
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
