% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftf.R
\name{createSafeFtF}
\alias{createSafeFtF}
\title{Removes disclosive variables}
\usage{
createSafeFtF(dt)
}
\description{
\code{createSafeFtF} removes the following disclosive variables:
 `Reg No`
 `Latitude`
 `Longitude`
 `Course (deg)`
}
\details{
Ideally these would be removed prior to sharing the data.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
