% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGridSpyData.R
\name{getCleanGridSpyData}
\alias{getCleanGridSpyData}
\title{Loads cleaned grid spy data into a data.table}
\usage{
getCleanGridSpyData(files)
}
\arguments{
\item{files}{the files to load (so you can pass a sunset if you wish)}

\item{pattern}{the circuit pattern to match}
}
\description{
\code{getCleanGridSpyData} loops over a file list and uses `readr::read_csv` to load and `rbind` them.
}
\details{
Use of read_csv enables .gz files to be autoloaded and proper parsing of dateTimes.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
