% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftf.R
\name{getEVMap}
\alias{getEVMap}
\title{Draws map of all locations (without sampling)}
\usage{
getEVMap(dt)
}
\description{
\code{getEVMap} Draws map pf all observations without sampling via the GPS Lat/Long. Observations without GPS co-ords (and thus time/date won't be mapped)
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
