% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGridSpyData.R
\name{getFileList}
\alias{getFileList}
\title{Get list of grid spy data files}
\usage{
getFileList(fpath, pattern)
}
\arguments{
\item{pattern}{the pattern to search within path}

\item{path}{the path to search}
}
\description{
\code{getFileList} takes a path and a pattern and recursively searches for data files in that path (and folders within it)
whose file names match the pattern.
}
\details{
Use the pattern to extract e.g : 1 min files (*at1.csv$) vs 1 second files etc

Returns the file list as a data.table with col name = `fList`. If no files are found returns a NULL dt
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
