% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGridSpyData.R
\name{getGridSpyFileList}
\alias{getGridSpyFileList}
\title{Get list of grid spy data files}
\usage{
getGridSpyFileList(fpath, pattern)
}
\arguments{
\item{pattern}{the pattern to search within path}

\item{path}{the path to search}
}
\description{
\code{getGridSpyFileList} takes a path and a pattern and searches for data files in that path which match the pattern.
Returns the file list as a data.frame. Use the pattern to extract 1 min files (*at1.csv$) vs 1 second files etc
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
