% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaData.R
\name{getMetaData}
\alias{getMetaData}
\title{Loads gridSpy meta data from .xlsx}
\usage{
getMetaData(f)
}
\arguments{
\item{f}{the .xlsx file}
}
\description{
\code{gertMetaData} does what it says. Expects .xlsx with two sheets, one for each sample.
}
\details{
Combines the results into one table
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
