% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftf.R
\name{inferLocationFtF}
\alias{inferLocationFtF}
\title{Infers coarse-grained location (home vs not-home)}
\usage{
inferLocationFtF(dt)
}
\description{
\code{inferLocationFtF} guesses 'home' by looking at location from 01:00 - 04:00. A future version might try geocoding other locations via lat/long look-up
}
\details{
adds `geoLoc` to the passed data.table and returns it
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
