% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Sets a range of parameters re-used throughout the repo}
\usage{
setup()
}
\description{
\code{setup} sets a range of parameters including data sources and .Rmd include locations. It returns these as a list.
}
\details{
The parameters can be over-written in scipts if needed but do so wih care!
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
