% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_tstrsplit.R
\name{testDTStringSplit}
\alias{testDTStringSplit}
\title{Tests data.table string split within a function}
\usage{
testDTStringSplit(dt)
}
\arguments{
\item{dt}{data.table with a column we wish to split & allocate to new columns}
}
\description{
\code{testDTStringSplit} splits the contents of a variable & allocates to new columns
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
