% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftf.R
\name{createDerivedFtF}
\alias{createDerivedFtF}
\title{Creates useful derived variables and returns dt}
\usage{
createDerivedFtF(dt)
}
\description{
\code{createDerivedFtF} converts the given time periods (TP1 -> TP48, 49. 50) to hh:mm. It ignores
 TP49 & TP50 as evil incarnations of DST related clock changes (see https://www.emi.ea.govt.nz/Wholesale/Datasets/Generation/Generation_MD/).
 We advise NEVER using the months in which this happens as it will hurt your brain.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
