% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGridSpyData.R
\name{getCleanGridSpyData}
\alias{getCleanGridSpyData}
\title{Loads cleaned grid spy power data for a given circuit between two dates into a data.table}
\usage{
getCleanGridSpyData(files, circuitPattern, dateFrom, dateTo)
}
\arguments{
\item{files}{the files to load (so you can pass a subset if you wish)}

\item{circuitPattern}{the circuit pattern to match}

\item{dateFrom}{date to start extract (inclusive)}

\item{dateTo}{date to end extract (inclusive)}
}
\description{
\code{getCleanGridSpyData} loops over a file list and loads each in turn using \code{readr::read_csv}. It filters each file to
extract data for particular circuits between two dates and creates some derived time/date variables before
using \code{rbind} to create a single data.table which is returned.
}
\details{
Function matches \code{circuitPattern} to extract specific circuits and selects observations between
 \code{dateFrom} and \code{dateTo}. Use this to extract any circuit you want between any given dates.

\code{circuitPattern} is passed to the \code{data.table} operator \code{\%like\%} so wild cards & stuff may work. YMMV

Use of \code{readr::read_csv} enables .gz files to be autoloaded and proper parsing of dateTimes.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk}
}
